Add files via upload
Database Connectivity Through mysql
Create a Java program with GUI to input students’ details such as index number, name,
names of the three subjects and the current GPA.
a. The details should be recorded in an SQLite database.
b. When user clicks on a button named as Retrieve, the name of all the students
should be shown in a list box.
c. Create another form to insert an index number and when user clicks on search,
the program should find the record relevant to that index number and show the
related details in a label.